Remove redundant conditional judgement in deletion
In a binary tree deletion, as its defination, first we need to find the
minimal node of its right sub-tree to replace its location if the node
to be deleted has both its left and right children. Then we shall trace
its sub-tree to find the minimal node in `ngx_rbtree_min`. Since the
terminal condition of the tracing loop has already been set as
`node->left == sentinel`, there is no need to leave the following
if-condition judging whether the minimal node's left child is the
sentinel any more because the code will never run here. And the program
will also run faster after removing the redundant conditional branch.

Signed-off-by: Leo Ma <begeekmyfriend@gmail.com>